Save correct device type in regression test
