+++ project/conftest.py
@@ -85,6 +85,8 @@ from hydra.core.hydra_config import HydraConfig
 from hydra_plugins.auto_schema import auto_schema_plugin
 from hydra_plugins.auto_schema.auto_schema_plugin import add_schemas_to_all_hydra_configs
 from omegaconf import DictConfig, open_dict
+from tensor_regression.stats import get_simple_attributes
+from tensor_regression.to_array import to_ndarray
 from torch import Tensor
 from torch.utils.data import DataLoader
 
@@ -678,3 +680,32 @@ def _patched_simple_attributes(v, precision: int | None):
     stats = tensor_regression.stats.get_simple_attributes(v, precision=precision)
     stats.pop("hash", None)
     return stats
+
+
+@get_simple_attributes.register(tuple)
+def _get_tuple_attributes(value: tuple, precision: int | None):
+    # This is called to get some simple stats to store in regression files during tests, in
+    # particular for tuples (since there isn't already a handler for it in the tensor_regression
+    # package.)
+    # Note: This information about this output is not very descriptive.
+    # not this is called only for the `out.past_key_values` entry in the `CausalLMOutputWithPast`
+    # that is returned from the forward pass output.
+    num_items_to_include = 5  # only show the stats of some of the items.
+    return {
+        "length": len(value),
+        **{
+            f"{i}": get_simple_attributes(item, precision=precision)
+            for i, item in enumerate(value[:num_items_to_include])
+        },
+    }
+
+
+@to_ndarray.register(list)
+@to_ndarray.register(tuple)
+def _tuple_to_ndarray(v: tuple | list):
+    """Convert a tuple of values to a numpy array to be stored in a regression file."""
+    # This could get a bit tricky because the items might not have the same shape and so on.
+    # However it seems like the ndarrays_regression fixture (which is what tensor_regression uses
+    # under the hood) is not complaining about us returning a list here, so we'll leave it at that
+    # for now.
+    return {i: to_ndarray(v_i) for i, v_i in enumerate(v)}  # type: ignore
